Delete compatibility.rb
It doesn't look like this is needed any more - (NB just doing this on the github UI to start a conversation), so the require would need removing and tests would need checking so obviously don't merge as is.

I had a look at the main fork of this project and due to time constraints decided not to try and upgrade that project to work with the latest devise, Rails and neo4j etc.

It looks like this is maybe nearly done though. Do you need any help with the last mile?